server/src/test/java/com/inti/nosily_coach/web/HelloControllerTest.java
package com.inti.nosily_coach.web;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.CoreMatchers.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

// 테스트를 진행할 때 JUnit에 내장된 실행자 외에 다른 실행자를 실행시킨다.
// SpringExtension라는 스프링 실행자를 사용한다.
// 즉, 스프링 부트 테스트와 JUnit 사이의 연결자 역할을 한다.
@ExtendWith(SpringExtension.class)
// 여러 스프링 테스트 어노테이션 중, Web에 집중할 수 있는 어노테이션이다.
// 선언할 경우 @Controller, @ControllerAdvice 등은 사용할 수 있지만,
// @Service, @Component, @Repository 등은 사용할 수 없다.
@WebMvcTest(controllers = HelloController.class)
public class HelloControllerTest {

    // 스프링이 관리하는 빈(Bean)을 주입 받는다.
    @Autowired
    // 웹 API를 테스트할 때 사용한다.
    // 스프링 MVC 테스트의 시작점
    // 이 클래스를 통해 HTTP GET, POST 등에 대한 API 테스트를 할 수 있다.
    private MockMvc mvc;

    @Test
    public void hello가_리턴된다() throws Exception {
        String hello = "Hello";

        // MockMvc를 통해 /hello 주소로 HTTP GET 요청
        // 체이닝이 되어 아래와 같이 여러 검증 기능을 이어서 선언할 수 있음
        mvc.perform(get("/hello"))
                // mvc.perform의 결과를 검증
                // HTTP Header의 Status를 검증 (200, 404, 500 ...)
                // 해당 코드에서는 OK, 즉 200인지를 검증
                .andExpect(status().isOk())
                // 응답 본문의 내용을 검증
                // 해당 코드에서는 리턴 값이 "Hello"가 맞는지 검증
                .andExpect(content().string(hello));
    }

    @Test
    public void helloDto가_리턴된다() throws Exception {
        String name = "hello";
        int amount = 1000;

        mvc.perform(
                        get("/hello/dto")
                                .param("name", name)
                                .param("amount", String.valueOf(amount))
                )
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.name", is(name)))
                .andExpect(jsonPath("$.amount", is(amount)));
    }
}